refactor: extract event fetching logic into useEvents hook
